Проект готов. Тестирование.

+++ ConsoleApp1/View/Pages/DeleteMedicamentPage.cs
@@ -8,28 +8,33 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace PharmacyApp.Pages
+namespace PharmacyApp.View.Pages
 {
-    public class DeleteMedicamentPage: IPage
+    public class DeleteMedicamentPage : IPage
     {
-    public string Create()
+        public string Create()
         {
             List<Medicament> list = new List<Medicament>();
             Medicament medicament = new Medicament();
-            IRepository<Medicament> medRepository = new MedicamentsRepository();
+            MedicamentsRepository medRepository = new MedicamentsRepository();
             list = medRepository.GetAll();
 
             Console.WriteLine(new string('-', 15 + 50 + 20 + 4));
             Console.WriteLine("|{0,15}|{1,50}|{2,20}|", "Id", "Name", "Price");
             Console.WriteLine(new string('-', 15 + 50 + 20 + 4));
 
-            foreach (Medicament med in list) 
+            foreach (Medicament med in list)
             {
-                Console.WriteLine("|{0,15}|{1,50}|{2,20}|",med.Id,med.Name,med.Price);
+                Console.WriteLine("|{0,15}|{1,50}|{2,20}|", med.Id, med.Name, med.Price);
             }
             Console.WriteLine(new string('-', 15 + 50 + 20 + 4));
-            Console.Write("Введите Id медикамента для удаления: ");
-            int id = Convert.ToInt32(Utils.GetValue(10));
+            Console.Write("Введите Id медикамента для удаления (для отмены операции нажмите ESC): ");
+
+            string result = Utils.GetValue(9);
+            if (result == "AbortOperation") return "Работа с товарами";
+            int id = Convert.ToInt32(result);
+
+
             if (medRepository.GetById(id).Count != 0)
             {
                 medRepository.Delete(id);
@@ -41,6 +46,7 @@ namespace PharmacyApp.Pages
                 Console.WriteLine("Медикамент с id={0} не найден", id);
                 Console.WriteLine("Для продолжения нажмите любую клавишу");
             }
+            Console.ReadKey();
             return "Работа с товарами";
         }
     }
