+++ ConsoleApp1/View/Pages/DeleteStorePage.cs
@@ -7,15 +7,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PharmacyApp.Pages
+namespace PharmacyApp.View.Pages
 {
-    public class DeleteConsignmentPage
+    public class DeleteStorePage : IPage
     {
-    public string Create()
+        public string Create()
         {
-            List<ConsignmentViewModel> list = new List<ConsignmentViewModel>();
+            List<Store> list = new List<Store>();
             StoresRepository storesRepository = new StoresRepository();
-           // list = storesRepository.GetAllAndPharmacyInformation();
+            list = storesRepository.GetAll();
 
             Console.WriteLine(new string('-', 15 + 25 + 25 + 40 + 4));
             Console.WriteLine("|{0,15}|{1,25}|{2,25}|{3,40}|", "Id", "Name", "PharmacyName", "PharmacyAddress");
@@ -23,14 +23,18 @@ namespace PharmacyApp.Pages
             var sortedList = from p in list
                              orderby p.Id
                              select p;
-          //  foreach (var store in sortedList)
-          //  {
-          //      Console.WriteLine("|{0,15}|{1,25}|{2,25}|{3,40}|", store.Id, store.Name, store.PharmacyName, store.PharmacyAddress);
-          //  }
+            foreach (var store in sortedList)
+            {
+                Console.WriteLine("|{0,15}|{1,25}|{2,25}|{3,40}|", store.Id, store.Name, store.PharmacyName, store.PharmacyAddress);
+            }
             Console.WriteLine(new string('-', 15 + 25 + 25 + 40 + 4));
 
-            Console.Write("Введите Id склада для удаления: ");
-            int id = Convert.ToInt32(Utils.GetValue(10));
+            Console.Write("Введите Id склада для удаления (для отмены операции нажмите ESC): ");
+
+            string result = Utils.GetValue(9);
+            if (result == "AbortOperation") return "Работа со складами";
+            int id = Convert.ToInt32(result);
+
             if (storesRepository.GetById(id).Count != 0)
             {
                 storesRepository.Delete(id);
@@ -42,7 +46,8 @@ namespace PharmacyApp.Pages
                 Console.WriteLine("Склад с id={0} не найден", id);
                 Console.WriteLine("Для продолжения нажмите любую клавишу");
             }
-            return "Работа с партиями";
+            Console.ReadKey();
+            return "Работа со складами";
         }
     }
 }
